+++ src/main/java/com/equipeturma862/cadastronf/service/NotaFiscalServiceImpl.java
@@ -1,11 +1,10 @@
 package com.equipeturma862.cadastronf.service;
 
-import com.equipeturma862.cadastronf.domain.ClassificacaoPessoa;
 import com.equipeturma862.cadastronf.domain.NotaFiscal;
 import com.equipeturma862.cadastronf.domain.Remetente;
 import com.equipeturma862.cadastronf.exceptions.NotaFiscalExists;
+import com.equipeturma862.cadastronf.exceptions.NotaFiscalNotFound;
 import com.equipeturma862.cadastronf.exceptions.RemetenteNotFound;
-import com.equipeturma862.cadastronf.exceptions.ValidationRequired;
 import com.equipeturma862.cadastronf.repository.NotasFiscaisRepositoy;
 import com.equipeturma862.cadastronf.repository.RemetenteRepositoy;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +12,6 @@ import org.apache.commons.collections4.IterableUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,18 +29,10 @@ public class NotaFiscalServiceImpl implements NotaFiscalService{
 
     @Override
     public NotaFiscal save(NotaFiscal notaFiscal, Long remetenteId) {
-        Optional<Remetente> remetente = remetenteRepositoy.findById(remetenteId);
-
-        if (!remetente.isEmpty()) {
-           if(remetente.get().getTipoDePessoa()==ClassificacaoPessoa.PESSOA_JURRIDICA &&
-                   (notaFiscal.getNumeroNotaFiscal() == null || notaFiscal.getTipoNaturezaOperacao() == null)) {
-                    throw new ValidationRequired();
-           }
-
+        if (remetenteRepositoy.existsById(remetenteId)) {
             if (notasFiscaisRepositoy.existsByNumeroNotaFiscal(notaFiscal.getNumeroNotaFiscal()) &&
-                    notasFiscaisRepositoy.existsByDataEmissao(notaFiscal.getDataEmissao()))
-            {
-                    throw new NotaFiscalExists();
+                    notasFiscaisRepositoy.existsByDataEmissao(notaFiscal.getDataEmissao())) {
+                throw new NotaFiscalExists();
             }else {
                     Remetente remetenteBuilder = Remetente.builder()
                             .id(remetenteId).build();
@@ -54,18 +44,25 @@ public class NotaFiscalServiceImpl implements NotaFiscalService{
 
     @Override
     public NotaFiscal getById(Long id) {
-        return notasFiscaisRepositoy.findById(id).get();
+        if (notasFiscaisRepositoy.existsById(id)) {
+            return notasFiscaisRepositoy.findById(id).get();
+        }
+        throw new NotaFiscalNotFound();
     }
 
     @Override
     public NotaFiscal update(Long id, NotaFiscal notaFiscal) {
-        notaFiscal.setId(id);
-        return notasFiscaisRepositoy.save(notaFiscal);
+        if (notasFiscaisRepositoy.existsById(id)) {
+            notaFiscal.setId(id);
+            return notasFiscaisRepositoy.save(notaFiscal);
+        }
+        throw new NotaFiscalNotFound();
     }
 
     @Override
     public void delete(Long id) {
-        notasFiscaisRepositoy.deleteById(id);
-
+        if (notasFiscaisRepositoy.existsById(id)) {
+            notasFiscaisRepositoy.deleteById(id);
+        } else {throw new NotaFiscalNotFound();}
     }
 }
